feat(arch): extract ManifoldModelCatalog from ManifoldInference kernel [P1d #1611]
Carves model discovery, catalog, benchmark, and image/video-gen record types
out of the ManifoldInference kernel into a new zero-dependency leaf product
`ManifoldModelCatalog` (depends: ManifoldHardware, ManifoldNetworking, ManifoldSecrets).

Source-compatibility is preserved by adding `@_exported import ManifoldModelCatalog`
in `Sources/ManifoldInference/ManifoldModelCatalogExport.swift` — every existing
`import ManifoldInference` call site resolves the moved symbols unchanged.

Files moved from ManifoldInference → ManifoldModelCatalog:
  Models/: APIEndpointRecord{,+Validation}, CloudModelManifestTable, DownloadState,
  Image{Generation{Config,ConfigSnapshot,Event},MessagePayload,ModelInfo,ModelLoadPlan,
  PlaceholderHash}, ModelDiscoveryError, ModelDownloadPlan, ModelInfo, ModelManifest,
  OperationalError, RemoteBackendConfig, SamplerPresetRecord,
  Video{GenerationConfig,GenerationConfigSnapshot,GenerationEvent,MessagePayload}
  Services/: BackgroundDownloadManaging, CloudBackendError, DiagnosticsService,
  HuggingFaceError, ModelCapabilityProbe, ModelCatalog, ModelStorageService
  Utilities/: Logging, UUID+v5
  Root: ManifoldConfiguration, ManifoldKitError
  SSEStreamLimits + SSEStreamError extracted from SSEStreamParser into new file
  (SSEStreamParser keeps GenerationEvent references that would create a cycle)

Kept in ManifoldInference (would require cascading kernel moves):
  SettingsService (uses ChatSessionRecord → kernel types)
  ModelBenchmarkRunner (uses InferenceService → @MainActor kernel)

Package.swift: new library product, target with ManifoldHardware/Networking/Secrets
deps, ManifoldModelCatalog added to ManifoldInference + ManifoldKit umbrella deps,
new ManifoldModelCatalogTests test target.

CI / scripts: ManifoldModelCatalogTests added to PROFILE_CI_XCTEST_FILTERS (test.sh)
and --filter chain (ci.yml). Affected-suites graph regenerated. Lint allowlist updated
for ManifoldModelCatalog/ModelInfo.swift UUID literal (same rationale as old path).
