+++ Sources/ManifoldModelCatalog/CloudModelManifestTable.swift
