+++ Sources/ManifoldModelCatalog/DiagnosticsService.swift
