+++ Sources/ManifoldModelCatalog/HuggingFaceError.swift
