+++ Sources/ManifoldModelCatalog/ImageGenerationConfig.swift
