+++ Sources/ManifoldModelCatalog/ImageGenerationConfigSnapshot.swift
