+++ Sources/ManifoldModelCatalog/ImageGenerationEvent.swift
