+++ Sources/ManifoldModelCatalog/ImageMessagePayload.swift
