+++ Sources/ManifoldModelCatalog/ImagePlaceholderHash.swift
