+++ Sources/ManifoldModelCatalog/ManifoldConfiguration.swift
