+++ Sources/ManifoldModelCatalog/ManifoldKitError.swift
