+++ Sources/ManifoldModelCatalog/ModelCapabilityProbe.swift
