+++ Sources/ManifoldModelCatalog/ModelDiscoveryError.swift
