+++ Sources/ManifoldModelCatalog/OperationalError.swift
