+++ Sources/ManifoldModelCatalog/SamplerPresetRecord.swift
