+++ Sources/ManifoldModelCatalog/UUID+v5.swift
