+++ Sources/ManifoldModelCatalog/VideoGenerationConfig.swift
