+++ Sources/ManifoldModelCatalog/VideoGenerationConfigSnapshot.swift
