+++ Sources/ManifoldModelCatalog/VideoGenerationEvent.swift
